Refactor server structure for maintainability
User prompt: The server code is becoming a bit unwieldy with everything in a single file. Please refactor it so that it is more manageable and easier to understand and update as we continue the project.
